+++ packages/netlify-cms-backend-github/src/GraphQLAPI.ts
@@ -72,6 +72,14 @@ type GraphQLPullRequest = {
     id: string;
     isFork: boolean;
   };
+  user: GraphQLPullsListResponseItemUser;
+};
+
+type GraphQLPullsListResponseItemUser = {
+  avatar_url: string;
+  login: string;
+  url: string;
+  name: string;
 };
 
 function transformPullRequest(pr: GraphQLPullRequest) {
@@ -270,6 +278,11 @@ export default class GraphQLAPI extends API {
     } else {
       return super.fetchBlobContent({ sha, repoURL, parseText });
     }
+  }
+
+  async getPullRequestAuthor(pullRequest: Octokit.PullsListResponseItem) {
+    const user = pullRequest.user as unknown as GraphQLPullsListResponseItemUser;
+    return user?.name || user?.login;
   }
 
   async getPullRequests(