+++ packages/netlify-cms-ui-default/src/Loader.js
@@ -121,7 +121,7 @@ const StyledLoader = styled(Loader)`
   position: absolute;
   top: 50%;
   left: 50%;
-  margin: 0px;
+  margin: 0;
   text-align: center;
   z-index: ${zIndex.zIndex1000};
   transform: translateX(-50%) translateY(-50%);
@@ -132,9 +132,9 @@ const StyledLoader = styled(Loader)`
     position: absolute;
     top: 0%;
     left: 50%;
-    width: 2.28571429rem;
-    height: 2.28571429rem;
-    margin: 0em 0em 0em -1.14285714rem;
+    width: 2.2857rem;
+    height: 2.2857rem;
+    margin: 0 0 0 -1.1429rem;
     border-radius: 500rem;
     border-style: solid;
     border-width: 0.2em;
@@ -150,7 +150,7 @@ const StyledLoader = styled(Loader)`
     animation: ${animations.loader} 0.6s linear;
     animation-iteration-count: infinite;
     border-color: ${colors.active} transparent transparent;
-    box-shadow: 0px 0px 0px 1px transparent;
+    box-shadow: 0 0 0 1px transparent;
   }
 `;