+++ packages/netlify-cms-widget-colorstring/src/ColorControl.js
@@ -70,10 +70,10 @@ const ColorPickerContainer = styled.div`
 // fullscreen div to close color picker when clicking outside of picker
 const ClickOutsideDiv = styled.div`
   position: fixed;
-  top: 0px;
-  right: 0px;
-  bottom: 0px;
-  left: 0px;
+  top: 0;
+  right: 0;
+  bottom: 0;
+  left: 0;
 `;
 
 export default class ColorControl extends React.Component {