+++ packages/netlify-cms/CHANGELOG.md
@@ -3,6 +3,198 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.10.188](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.10.187...netlify-cms@2.10.188) (2022-03-08)
+
+**Note:** Version bump only for package netlify-cms
+
+
+
+
+
+## [2.10.187](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.10.186...netlify-cms@2.10.187) (2022-02-28)
+
+**Note:** Version bump only for package netlify-cms
+
+
+
+
+
+## [2.10.186](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.10.185...netlify-cms@2.10.186) (2022-02-08)
+
+**Note:** Version bump only for package netlify-cms
+
+
+
+
+
+## [2.10.185](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.10.184...netlify-cms@2.10.185) (2022-01-21)
+
+**Note:** Version bump only for package netlify-cms
+
+
+
+
+
+## [2.10.184](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.10.183...netlify-cms@2.10.184) (2022-01-21)
+
+**Note:** Version bump only for package netlify-cms
+
+
+
+
+
+## [2.10.183](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.10.182...netlify-cms@2.10.183) (2021-12-28)
+
+**Note:** Version bump only for package netlify-cms
+
+
+
+
+
+## [2.10.182](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.10.181...netlify-cms@2.10.182) (2021-12-21)
+
+**Note:** Version bump only for package netlify-cms
+
+
+
+
+
+## [2.10.181](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.10.180...netlify-cms@2.10.181) (2021-12-08)
+
+**Note:** Version bump only for package netlify-cms
+
+
+
+
+
+## [2.10.180](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.10.179...netlify-cms@2.10.180) (2021-12-03)
+
+**Note:** Version bump only for package netlify-cms
+
+
+
+
+
+## [2.10.179](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.10.178...netlify-cms@2.10.179) (2021-11-12)
+
+**Note:** Version bump only for package netlify-cms
+
+
+
+
+
+## [2.10.178](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.10.177...netlify-cms@2.10.178) (2021-11-01)
+
+**Note:** Version bump only for package netlify-cms
+
+
+
+
+
+## [2.10.177](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.10.176...netlify-cms@2.10.177) (2021-10-28)
+
+**Note:** Version bump only for package netlify-cms
+
+
+
+
+
+## [2.10.176](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.10.175...netlify-cms@2.10.176) (2021-10-28)
+
+**Note:** Version bump only for package netlify-cms
+
+
+
+
+
+## [2.10.175](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.10.174...netlify-cms@2.10.175) (2021-10-25)
+
+**Note:** Version bump only for package netlify-cms
+
+
+
+
+
+## [2.10.174](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.10.173...netlify-cms@2.10.174) (2021-10-21)
+
+**Note:** Version bump only for package netlify-cms
+
+
+
+
+
+## [2.10.173](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.10.172...netlify-cms@2.10.173) (2021-10-18)
+
+**Note:** Version bump only for package netlify-cms
+
+
+
+
+
+## [2.10.172](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.10.171...netlify-cms@2.10.172) (2021-10-15)
+
+**Note:** Version bump only for package netlify-cms
+
+
+
+
+
+## [2.10.171](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.10.170...netlify-cms@2.10.171) (2021-10-11)
+
+**Note:** Version bump only for package netlify-cms
+
+
+
+
+
+## [2.10.170](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.10.169...netlify-cms@2.10.170) (2021-10-11)
+
+**Note:** Version bump only for package netlify-cms
+
+
+
+
+
+## [2.10.169](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.10.168...netlify-cms@2.10.169) (2021-10-08)
+
+**Note:** Version bump only for package netlify-cms
+
+
+
+
+
+## [2.10.168](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.10.167...netlify-cms@2.10.168) (2021-10-07)
+
+**Note:** Version bump only for package netlify-cms
+
+
+
+
+
+## [2.10.167](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.10.166...netlify-cms@2.10.167) (2021-10-07)
+
+**Note:** Version bump only for package netlify-cms
+
+
+
+
+
+## [2.10.166](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.10.165...netlify-cms@2.10.166) (2021-09-30)
+
+**Note:** Version bump only for package netlify-cms
+
+
+
+
+
+## [2.10.165](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.10.164...netlify-cms@2.10.165) (2021-09-29)
+
+**Note:** Version bump only for package netlify-cms
+
+
+
+
+
 ## [2.10.164](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.10.163...netlify-cms@2.10.164) (2021-09-13)
 
 **Note:** Version bump only for package netlify-cms